Adds .gitignore and small changes
adds a .gitignore file and some small changes to the SubHunter.java file.
need to find a good 'best practice' sort of outline on what are the minimal
files needed for an effective android repository. Probably don't need the res
folder or the idea folder or the build folder unless it's something I have
specifically changed. :thinking_face: